+++ CHANGES.md
@@ -1,3 +1,3 @@
 0.2.0
-    - changed deriveSchema to work with PostgreSQL >= 11.0
+    - updated deriveSchema to work with PostgreSQL >= 11.0
     - updated packages to remove security vulnerabilities